refactor #158: Payment 도메인 패키지 변경
